Move `SNPs` tests to `test_snps.py`
