+++ tests/test_snps.py
@@ -39,7 +39,11 @@ import gzip
 import pandas as pd
 import shutil
 
+import numpy as np
+
 from snps import SNPs
+from snps.utils import create_dir
+from snps.resources import Resources, ReferenceSequence
 from tests import BaseSNPsTestCase
 
 
@@ -224,3 +228,623 @@ class TestSnps(BaseSNPsTestCase):
         assert s.snp_count == 0
         assert list(s.snps.columns.values) == ["chrom", "pos", "genotype"]
         assert s.snps.index.name == "rsid"
+
+    def generic_snps(self):
+        return self.create_snp_df(
+            rsid=["rs" + str(i) for i in range(1, 9)],
+            chrom=["1"] * 8,
+            pos=list(range(101, 109)),
+            genotype=["AA", "CC", "GG", "TT", np.nan, "GC", "TC", "AT"],
+        )
+
+    def generic_snps_vcf(self):
+        df = self.generic_snps()
+        return df.append(
+            self.create_snp_df(
+                rsid=["rs" + str(i) for i in range(12, 18)],
+                chrom=["1"] * 6,
+                pos=list(range(112, 118)),
+                genotype=[np.nan] * 6,
+            )
+        )
+
+    def snps_GRCh38_func(self):
+        return self.create_snp_df(
+            rsid=["rs3094315", "rsIndelTest", "rs2500347", "rs11928389"],
+            chrom=["1", "1", "1", "3"],
+            pos=[817186, 148946168, 148946169, 50889578],
+            genotype=["AA", "ID", np.nan, "TC"],
+        )
+
+    def snps_GRCh38_PAR(self):
+        return self.create_snp_df(
+            rsid=["rs28736870", "rs113378274", "rs113313554"],
+            chrom=["X", "X", "Y"],
+            pos=[304103, 93431058, 624523],
+            genotype=["AA", "AA", "AA"],
+        )
+
+    def test_snps_DNALand(self):
+        # https://dna.land/
+        s = SNPs("tests/input/DNALand.txt")
+        assert s.source == "DNA.Land"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+    def test_snps_23andme(self):
+        # https://www.23andme.com
+        s = SNPs("tests/input/23andme.txt")
+        assert s.source == "23andMe"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+    def test_snps_23andme_zip(self):
+        with atomic_write(
+            "tests/input/23andme.txt.zip", mode="wb", overwrite=True
+        ) as f:
+            with zipfile.ZipFile(f, "w") as f_zip:
+                # https://stackoverflow.com/a/16104667
+                f_zip.write("tests/input/23andme.txt", arcname="23andme.txt")
+        s = SNPs("tests/input/23andme.txt.zip")
+        assert s.source == "23andMe"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+    def test_snps_ftdna(self):
+        # https://www.familytreedna.com
+        s = SNPs("tests/input/ftdna.csv")
+        assert s.source == "FTDNA"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+    def test_snps_ftdna_gzip(self):
+        with open("tests/input/ftdna.csv", "rb") as f_in:
+            with atomic_write(
+                "tests/input/ftdna.csv.gz", mode="wb", overwrite=True
+            ) as f_out:
+                with gzip.open(f_out, "wb") as f_gzip:
+                    shutil.copyfileobj(f_in, f_gzip)
+        s = SNPs("tests/input/ftdna.csv.gz")
+        assert s.source == "FTDNA"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+    def test_snps_ftdna_famfinder(self):
+        # https://www.familytreedna.com
+        s = SNPs("tests/input/ftdna_famfinder.csv")
+        assert s.source == "FTDNA"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+    def test_snps_ancestry(self):
+        # https://www.ancestry.com
+        s = SNPs("tests/input/ancestry.txt")
+        assert s.source == "AncestryDNA"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+    def test_snps_genes_for_good(self):
+        # https://genesforgood.sph.umich.edu/
+        s = SNPs("tests/input/genesforgood.txt")
+        assert s.source == "GenesForGood"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+    def test_snps_myheritage(self):
+        # https://www.myheritage.com
+        s = SNPs("tests/input/myheritage.csv")
+        assert s.source == "MyHeritage"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+    @staticmethod
+    def _setup_gsa_test():
+        # reset resource if already loaded
+        temp = SNPs()
+        temp._resources._gsa_resources = {}
+
+        create_dir("resources")
+
+        with open("tests/resources/gsa_rsid_map.txt", "rb") as f_in:
+            with atomic_write(
+                "resources/gsa_rsid_map.txt.gz", mode="wb", overwrite=True
+            ) as f_out:
+                with gzip.open(f_out, "wb") as f_gzip:
+                    shutil.copyfileobj(f_in, f_gzip)
+
+        with open("tests/resources/gsa_chrpos_map.txt", "rb") as f_in:
+            with atomic_write(
+                "resources/gsa_chrpos_map.txt.gz", mode="wb", overwrite=True
+            ) as f_out:
+                with gzip.open(f_out, "wb") as f_gzip:
+                    shutil.copyfileobj(f_in, f_gzip)
+
+    @staticmethod
+    def _teardown_gsa_test():
+        os.remove("resources/gsa_rsid_map.txt.gz")
+        os.remove("resources/gsa_chrpos_map.txt.gz")
+
+    def test_snps_codigo46_bytes(self):
+        # https://codigo46.com.mx
+
+        self._setup_gsa_test()
+
+        with open("tests/input/codigo46.txt", "rb") as f:
+            s = SNPs(f.read())
+        assert s.source == "Codigo46"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+        self._teardown_gsa_test()
+
+    def test_snps_codigo46(self):
+        # https://codigo46.com.mx
+
+        self._setup_gsa_test()
+
+        s = SNPs("tests/input/codigo46.txt")
+        assert s.source == "Codigo46"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+        self._teardown_gsa_test()
+
+    def test_snps_sano_bytes(self):
+        # https://sanogenetics.com
+
+        self._setup_gsa_test()
+
+        with open("tests/input/sano.txt", "rb") as f:
+            s = SNPs(f.read())
+        assert s.source == "Sano"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+        self._teardown_gsa_test()
+
+    def test_snps_sano(self):
+        # https://sanogenetics.com
+
+        self._setup_gsa_test()
+
+        s = SNPs("tests/input/sano.txt")
+        assert s.source == "Sano"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+        self._teardown_gsa_test()
+
+    def test_snps_livingdna(self):
+        # https://livingdna.com
+        s = SNPs("tests/input/livingdna.csv")
+        assert s.source == "LivingDNA"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+    def test_snps_mapmygenome(self):
+        # https://mapmygenome.in
+        s = SNPs("tests/input/mapmygenome.txt")
+        assert s.source == "Mapmygenome"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps())
+
+    def test_snps_vcf(self):
+        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
+        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
+        # phased snps, and snps with missing rsID
+        s = SNPs("tests/input/testvcf.vcf")
+        assert s.source == "vcf"
+        assert not s.unannotated_vcf
+        assert not s.phased
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
+
+    def test_snps_vcf_phased(self):
+        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
+        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
+        # phased snps, and snps with missing rsID
+        s = SNPs("tests/input/testvcf_phased.vcf")
+        assert s.source == "vcf"
+        assert not s.unannotated_vcf
+        assert s.phased
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
+
+    def test_snps_vcf_rsids(self):
+        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
+        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
+        # phased snps, and snps with missing rsID
+        rsids = ["rs1", "rs2"]
+        s = SNPs("tests/input/testvcf.vcf", rsids=rsids)
+        assert s.source == "vcf"
+        assert not s.unannotated_vcf
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf().loc[rsids])
+
+    def test_snps_vcf_gz(self):
+        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
+        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
+        # phased snps, and snps with missing rsID
+        with open("tests/input/testvcf.vcf", "rb") as f_in:
+            with atomic_write(
+                "tests/input/testvcf.vcf.gz", mode="wb", overwrite=True
+            ) as f_out:
+                with gzip.open(f_out, "wb") as f_gzip:
+                    shutil.copyfileobj(f_in, f_gzip)
+
+        s = SNPs("tests/input/testvcf.vcf.gz")
+        assert s.source == "vcf"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
+
+    def test_snps_vcf_gz_rsids(self):
+        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
+        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
+        # phased snps, and snps with missing rsID
+        with open("tests/input/testvcf.vcf", "rb") as f_in:
+            with atomic_write(
+                "tests/input/testvcf.vcf.gz", mode="wb", overwrite=True
+            ) as f_out:
+                with gzip.open(f_out, "wb") as f_gzip:
+                    shutil.copyfileobj(f_in, f_gzip)
+
+        rsids = ["rs1", "rs2"]
+        s = SNPs("tests/input/testvcf.vcf.gz", rsids=rsids)
+        assert s.source == "vcf"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf().loc[rsids])
+
+    def test_snps_unannotated_vcf(self):
+        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
+        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
+        # phased snps, and snps with missing rsID
+        s = SNPs("tests/input/unannotated_testvcf.vcf")
+        assert s.source == "vcf"
+        assert s.unannotated_vcf
+
+    def test_snps_not_phased(self):
+        s = SNPs("tests/input/generic.csv")
+        assert s.source == "generic"
+        assert not s.phased
+
+    def test_snps_vcf_buffer(self):
+        with open("tests/input/testvcf.vcf", "r") as f:
+            snps_vcf_buffer = SNPs(f.read().encode("utf-8"))
+        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
+        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
+        # phased snps, and snps with missing rsID
+        assert snps_vcf_buffer.source == "vcf"
+        pd.testing.assert_frame_equal(snps_vcf_buffer.snps, self.generic_snps_vcf())
+
+    def test_snps_vcf_buffer_rsids(self):
+        with open("tests/input/testvcf.vcf", "r") as f:
+            rsids = ["rs1", "rs2"]
+            df = SNPs(f.read().encode("utf-8"), rsids=rsids)
+        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
+        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
+        # phased snps, and snps with missing rsID
+        assert df.source == "vcf"
+        pd.testing.assert_frame_equal(df.snps, self.generic_snps_vcf().loc[rsids])
+
+    def test_snps_vcf_buffer_gz(self):
+        with open("tests/input/testvcf.vcf", "rb") as f_in:
+            with atomic_write(
+                "tests/input/testvcf.vcf.gz", mode="wb", overwrite=True
+            ) as f_out:
+                with gzip.open(f_out, "wb") as f_gzip:
+                    shutil.copyfileobj(f_in, f_gzip)
+
+        with open("tests/input/testvcf.vcf.gz", "rb") as f:
+            data = f.read()
+            s = SNPs(data)
+        os.remove("tests/input/testvcf.vcf.gz")
+        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
+        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
+        # phased snps, and snps with missing rsID
+        assert s.source == "vcf"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
+
+    def test_snps_vcf_buffer_gz_rsids(self):
+        with open("tests/input/testvcf.vcf", "rb") as f_in:
+            with atomic_write(
+                "tests/input/testvcf.vcf.gz", mode="wb", overwrite=True
+            ) as f_out:
+                with gzip.open(f_out, "wb") as f_gzip:
+                    shutil.copyfileobj(f_in, f_gzip)
+
+        with open("tests/input/testvcf.vcf.gz", "rb") as f:
+            rsids = ["rs1", "rs2"]
+            data = f.read()
+            s = SNPs(data, rsids=rsids)
+        os.remove("tests/input/testvcf.vcf.gz")
+
+        # https://samtools.github.io/hts-specs/VCFv4.2.pdf
+        # this tests for homozygous snps, heterozygous snps, multiallelic snps,
+        # phased snps, and snps with missing rsID
+        assert s.source == "vcf"
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf().loc[rsids])
+
+    def test_source_generic(self):
+        s = SNPs("tests/input/NCBI36.csv")
+        assert s.source == "generic"
+
+    def test_snps_no_snps(self):
+        s = SNPs()
+        assert s.snps.empty
+
+    # def test_snp_count(self):
+    #     s = SNPs("tests/input/NCBI36.csv")
+    #     assert s.snp_count == 4
+    #
+    # def test_snp_count_no_snps(self):
+    #     s = SNPs()
+    #     assert s.snp_count == 0
+    #
+    # def test_chromosomes(self):
+    #     s = SNPs("tests/input/chromosomes.csv")
+    #     assert s.chromosomes == ["1", "2", "3", "5", "PAR", "MT"]
+    #
+    # def test_chromosomes_no_snps(self):
+    #     s = SNPs()
+    #     assert s.chromosomes == []
+    #
+    # def test_chromosomes_summary(self):
+    #     s = SNPs("tests/input/chromosomes.csv")
+    #     assert s.chromosomes_summary == "1-3, 5, PAR, MT"
+    #
+    # def test_chromosomes_summary_no_snps(self):
+    #     s = SNPs()
+    #     assert s.chromosomes_summary == ""
+
+    def test_build(self):
+        s = SNPs("tests/input/NCBI36.csv")
+        assert s.build == 36
+        assert s.assembly == "NCBI36"
+
+    # def test_sex_Male_Y_chrom(self):
+    #     s = self.simulate_snps(chrom="Y", pos_start=1, pos_max=59373566, pos_step=10000)
+    #     assert s.sex == "Male"
+
+    def test_sex_Female_Y_chrom(self):
+        s = self.simulate_snps(
+            chrom="Y", pos_start=1, pos_max=59373566, pos_step=10000, null_snp_step=1
+        )
+        assert s.sex == "Female"
+
+    def test_sex_Female_X_chrom(self):
+        s = self.simulate_snps(
+            chrom="X", pos_start=1, pos_max=155270560, pos_step=10000, genotype="AC"
+        )
+        assert s.sex == "Female"
+
+    def test_sex_Male_X_chrom(self):
+        s = self.simulate_snps(
+            chrom="X", pos_start=1, pos_max=155270560, pos_step=10000, genotype="AA"
+        )
+        assert s.snp_count == 15528
+        s._deduplicate_XY_chrom()
+        assert s.snp_count == 15528
+        assert len(s.discrepant_XY_snps) == 0
+        assert s.sex == "Male"
+
+    def test_sex_Male_X_chrom_discrepant_XY_snps(self):
+        s = self.simulate_snps(
+            chrom="X", pos_start=1, pos_max=155270560, pos_step=10000, genotype="AA"
+        )
+        assert s.snp_count == 15528
+        s._snps.loc["rs8001", "genotype"] = "AC"
+        s._deduplicate_XY_chrom()
+        assert s.snp_count == 15527
+        result = self.create_snp_df(
+            rsid=["rs8001"], chrom=["X"], pos=[80000001], genotype=["AC"]
+        )
+        pd.testing.assert_frame_equal(s.discrepant_XY_snps, result)
+        assert s.sex == "Male"
+
+    def test_sex_not_determined(self):
+        s = self.simulate_snps(
+            chrom="1", pos_start=1, pos_max=249250621, pos_step=10000
+        )
+        assert s.sex == ""
+
+    def test_save_snps(self):
+        snps = SNPs("tests/input/GRCh37.csv")
+        assert os.path.relpath(snps.save_snps()) == "output/generic_GRCh37.csv"
+        s_saved = SNPs("output/generic_GRCh37.csv")
+        pd.testing.assert_frame_equal(s_saved.snps, self.snps_GRCh37())
+
+    def test_save_snps_bytes(self):
+        snps = SNPs("tests/input/GRCh37.csv")
+        assert os.path.relpath(snps.save_snps()) == "output/generic_GRCh37.csv"
+        with open("output/generic_GRCh37.csv", "rb") as f:
+            s_saved = SNPs(f.read())
+        pd.testing.assert_frame_equal(s_saved.snps, self.snps_GRCh37())
+
+    def test_save_snps_tsv(self):
+        snps = SNPs("tests/input/generic.csv")
+        assert (
+            os.path.relpath(snps.save_snps("generic.tsv", sep="\t"))
+            == "output/generic.tsv"
+        )
+        s_saved = SNPs("output/generic.tsv")
+        assert snps.source == "generic"
+        pd.testing.assert_frame_equal(s_saved.snps, self.generic_snps())
+
+    def test_save_snps_tsv_bytes(self):
+        snps = SNPs("tests/input/generic.csv")
+        assert (
+            os.path.relpath(snps.save_snps("generic.tsv", sep="\t"))
+            == "output/generic.tsv"
+        )
+        with open("output/generic.tsv", "rb") as f:
+            s_saved = SNPs(f.read())
+        assert snps.source == "generic"
+        pd.testing.assert_frame_equal(s_saved.snps, self.generic_snps())
+
+    def test_save_snps_vcf(self):
+        s = SNPs("tests/input/testvcf.vcf")
+
+        r = Resources()
+        r._reference_sequences["GRCh37"] = {}
+        with open("tests/input/generic.fa", "rb") as f_in:
+            with atomic_write(
+                "tests/input/generic.fa.gz", mode="wb", overwrite=True
+            ) as f_out:
+                with gzip.open(f_out, "wb") as f_gzip:
+                    shutil.copyfileobj(f_in, f_gzip)
+
+        seq = ReferenceSequence(ID="1", path="tests/input/generic.fa.gz")
+
+        r._reference_sequences["GRCh37"]["1"] = seq
+
+        assert os.path.relpath(s.save_snps(vcf=True)) == "output/vcf_GRCh37.vcf"
+        s = SNPs("output/vcf_GRCh37.vcf")
+        assert not s.phased
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
+
+    def test_save_snps_vcf_phased(self):
+        # read phased data
+        s = SNPs("tests/input/testvcf_phased.vcf")
+
+        # setup resource to use test FASTA reference sequence
+        r = Resources()
+        r._reference_sequences["GRCh37"] = {}
+        with open("tests/input/generic.fa", "rb") as f_in:
+            with atomic_write(
+                "tests/input/generic.fa.gz", mode="wb", overwrite=True
+            ) as f_out:
+                with gzip.open(f_out, "wb") as f_gzip:
+                    shutil.copyfileobj(f_in, f_gzip)
+
+        seq = ReferenceSequence(ID="1", path="tests/input/generic.fa.gz")
+
+        r._reference_sequences["GRCh37"]["1"] = seq
+
+        # save phased data to VCF
+        assert os.path.relpath(s.save_snps(vcf=True)) == "output/vcf_GRCh37.vcf"
+        # read saved VCF
+        s = SNPs("output/vcf_GRCh37.vcf")
+        assert s.phased
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
+
+    def test_save_snps_csv_phased(self):
+        # read phased data
+        s = SNPs("tests/input/testvcf_phased.vcf")
+        # save phased data to CSV
+        assert os.path.relpath(s.save_snps()) == "output/vcf_GRCh37.csv"
+        # read saved CSV
+        s = SNPs("output/vcf_GRCh37.csv")
+        assert s.phased
+        pd.testing.assert_frame_equal(s.snps, self.generic_snps_vcf())
+
+    def test_save_snps_specify_file(self):
+        s = SNPs("tests/input/GRCh37.csv")
+        assert os.path.relpath(s.save_snps("snps.csv")) == "output/snps.csv"
+        s_saved = SNPs("output/snps.csv")
+        pd.testing.assert_frame_equal(s_saved.snps, self.snps_GRCh37())
+
+    def test_save_snps_no_snps(self):
+        s = SNPs()
+        assert not s.save_snps()
+
+    def test_save_snps_no_snps_vcf(self):
+        s = SNPs()
+        assert not s.save_snps(vcf=True)
+
+    def test_remap_snps_36_to_37(self):
+        s = SNPs("tests/input/NCBI36.csv")
+        chromosomes_remapped, chromosomes_not_remapped = s.remap_snps(37)
+        assert s.build == 37
+        assert s.assembly == "GRCh37"
+        assert len(chromosomes_remapped) == 2
+        assert len(chromosomes_not_remapped) == 0
+        pd.testing.assert_frame_equal(s.snps, self.snps_GRCh37())
+
+    def test_remap_snps_36_to_37_multiprocessing(self):
+        s = SNPs("tests/input/NCBI36.csv", parallelize=True)
+        chromosomes_remapped, chromosomes_not_remapped = s.remap_snps(37)
+        assert s.build == 37
+        assert s.assembly == "GRCh37"
+        assert len(chromosomes_remapped) == 2
+        assert len(chromosomes_not_remapped) == 0
+        pd.testing.assert_frame_equal(s.snps, self.snps_GRCh37())
+
+    def test_remap_snps_37_to_36(self):
+        s = SNPs("tests/input/GRCh37.csv")
+        chromosomes_remapped, chromosomes_not_remapped = s.remap_snps(36)
+        assert s.build == 36
+        assert s.assembly == "NCBI36"
+        assert len(chromosomes_remapped) == 2
+        assert len(chromosomes_not_remapped) == 0
+        pd.testing.assert_frame_equal(s.snps, self.snps_NCBI36())
+
+    def test_remap_snps_37_to_38(self):
+        s = SNPs("tests/input/GRCh37.csv")
+        chromosomes_remapped, chromosomes_not_remapped = s.remap_snps(38)
+        assert s.build == 38
+        assert s.assembly == "GRCh38"
+        assert len(chromosomes_remapped) == 2
+        assert len(chromosomes_not_remapped) == 0
+        pd.testing.assert_frame_equal(s.snps, self.snps_GRCh38_func())
+
+    def test_remap_snps_37_to_38_with_PAR_SNP(self):
+        if (
+            not os.getenv("DOWNLOADS_ENABLED")
+            or os.getenv("DOWNLOADS_ENABLED") == "true"
+        ):
+            s = SNPs("tests/input/GRCh37_PAR.csv")
+            assert s.snp_count == 4
+            chromosomes_remapped, chromosomes_not_remapped = s.remap_snps(38)
+            assert s.build == 38
+            assert s.assembly == "GRCh38"
+            assert len(chromosomes_remapped) == 2
+            assert len(chromosomes_not_remapped) == 1
+            assert s.snp_count == 3
+            pd.testing.assert_frame_equal(s.snps, self.snps_GRCh38_PAR())
+
+    def test_remap_snps_37_to_37(self):
+        s = SNPs("tests/input/GRCh37.csv")
+        chromosomes_remapped, chromosomes_not_remapped = s.remap_snps(37)
+        assert s.build == 37
+        assert s.assembly == "GRCh37"
+        assert len(chromosomes_remapped) == 0
+        assert len(chromosomes_not_remapped) == 2
+        pd.testing.assert_frame_equal(s.snps, self.snps_GRCh37())
+
+    def test_remap_snps_no_snps(self):
+        s = SNPs()
+        chromosomes_remapped, chromosomes_not_remapped = s.remap_snps(38)
+        assert not s.build
+        assert len(chromosomes_remapped) == 0
+        assert len(chromosomes_not_remapped) == 0
+
+    def test_remap_snps_invalid_assembly(self):
+        s = SNPs("tests/input/GRCh37.csv")
+        chromosomes_remapped, chromosomes_not_remapped = s.remap_snps(-1)
+        assert s.build == 37
+        assert s.assembly == "GRCh37"
+        assert len(chromosomes_remapped) == 0
+        assert len(chromosomes_not_remapped) == 2
+
+    def test___repr__snps(self):
+        s = SNPs("tests/input/GRCh37.csv")
+        assert "SNPs('tests/input/GRCh37.csv')" == s.__repr__()
+
+    def test_load_opensnp_datadump_file(self):
+        # temporarily set resources dir to tests
+        r = Resources()
+        r._resources_dir = "tests/resources"
+
+        # write test openSNP datadump zip
+        with atomic_write(
+            "tests/resources/opensnp_datadump.current.zip", mode="wb", overwrite=True
+        ) as f:
+            with zipfile.ZipFile(f, "w") as f_zip:
+                f_zip.write("tests/input/generic.csv", arcname="generic1.csv")
+                f_zip.write("tests/input/generic.csv", arcname="generic2.csv")
+
+        snps1 = SNPs(r.load_opensnp_datadump_file("generic1.csv"))
+        snps2 = SNPs(r.load_opensnp_datadump_file("generic2.csv"))
+
+        pd.testing.assert_frame_equal(snps1.snps, self.generic_snps())
+        pd.testing.assert_frame_equal(snps2.snps, self.generic_snps())
+
+        r._resources_dir = "resources"
+
+    def test_heterozygous_snps(self):
+        s = SNPs("tests/input/generic.csv")
+        pd.testing.assert_frame_equal(
+            s.heterozygous_snps(),
+            self.create_snp_df(
+                rsid=["rs6", "rs7", "rs8"],
+                chrom=["1", "1", "1"],
+                pos=[106, 107, 108],
+                genotype=["GC", "TC", "AT"],
+            ),
+        )
+
+    def test_not_null_snps(self):
+        s = SNPs("tests/input/generic.csv")
+        snps = self.generic_snps()
+        snps.drop("rs5", inplace=True)
+        pd.testing.assert_frame_equal(s.not_null_snps(), snps)
